Created custom hook to track editor state
REF https://linear.app/ghost/issue/PLG-298/prevent-top-level-comment-input-from-closing-when-it-has-text
- Extracting the editor state allows us to remove duplicate logic from Form and MainForm. It also centralises editor behaviour which will help a future switch over to Koenig.
