+++ apps/comments-ui/src/components/content/forms/MainForm.tsx
@@ -1,9 +1,10 @@
-import React, {useCallback, useEffect, useRef, useState} from 'react';
+import React, {useCallback, useEffect, useRef} from 'react';
 import {Form, FormWrapper} from './Form';
 import {getEditorConfig} from '../../../utils/editor';
 import {scrollToElement} from '../../../utils/helpers';
 import {useAppContext} from '../../../AppContext';
 import {useEditor} from '@tiptap/react';
+import {useEditorState} from '../../../utils/hooks';
 
 type Props = {
     commentsCount: number
@@ -11,32 +12,17 @@ type Props = {
 
 const MainForm: React.FC<Props> = ({commentsCount}) => {
     const {postId, dispatchAction, t} = useAppContext();
-    const [hasContent, setHasContent] = useState(false);
 
     const config = {
         placeholder: (commentsCount === 0 ? t('Start the conversation') : t('Join the discussion')),
         autofocus: false
     };
 
-    const editor = useEditor({
+    const baseEditor = useEditor({
         ...getEditorConfig(config)
     }, [commentsCount]);
 
-    useEffect(() => {
-        if (editor) {
-            const checkContent = () => {
-                setHasContent(!editor.isEmpty);
-            };
-            
-            editor.on('update', checkContent);
-            editor.on('transaction', checkContent);
-
-            return () => {
-                editor.off('update', checkContent);
-                editor.off('transaction', checkContent);
-            };
-        }
-    }, [editor]);
+    const {editor, hasContent} = useEditorState(baseEditor);
 
     const submit = useCallback(async ({html}) => {
         // Send comment to server
